requirements.txt
django>=4.2,<7.0
djangorestframework>=3.14
httpx>=0.25
